+++ ts/packages/marginfi-cli/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.7](https://github.com/mrgnlabs/marginfi-sdk/compare/@mrgnlabs/marginfi-cli@0.5.6...@mrgnlabs/marginfi-cli@0.5.7) (2022-09-14)
+
+**Note:** Version bump only for package @mrgnlabs/marginfi-cli
+
+
+
+
+
 ## [0.5.2](https://github.com/mrgnlabs/marginfi-sdk/compare/@mrgnlabs/marginfi-cli@0.5.1...@mrgnlabs/marginfi-cli@0.5.2) (2022-09-06)
 
 **Note:** Version bump only for package @mrgnlabs/marginfi-cli